Add a target to randomize the hostname
I'm not 100% sure I'm going to use this, but I wrote it so I'll leave it
here for now. The idea is if I have multiple machines running NixOS, I'd
probably want them to have a unique hostname. I might be able to use
something like the hardware name, i.e., `neofetch` reports this host to
be `LENOVO 20R1000RUS`, so maybe I could use the last half of that?

Something to think about at least.